fix build : change int i -> size_t i

+++ src/applications/gaming_app/ReportListView.cpp
@@ -178,7 +178,7 @@ void ReportListView::NotifyUpdated( const Reports& reports )
     tools::Iterator< const Report& > iterator = reports.CreateIterator();
     while( iterator.HasMoreElements() )
         reportModel_.appendRow( iterator.NextElement().GetReportData() );
-    for( int i = 0; i < reports.traces_.size(); ++i )
+    for( size_t i = 0; i < reports.traces_.size(); ++i )
         reportModel_.appendRow( reports.traces_[ i ]->GetReportData() );
 }
